\name{front41Data}
\docType{data}
\alias{front41Data}

\title{Data provided with Tim Coelli's Frontier 4.1}

\description{
   The \code{front41Data} data frame contains cross-sectional data of 60 firms.
}

\usage{data(front41Data)}

\format{
   This data frame contains the following columns:
   \describe{
      \item{firm}{firm ID.}
      \item{output}{output.}
      \item{capital}{capital.}
      \item{labour}{labour.}
   }
}

\source{
   Coelli, T. (1996) A Guide to FRONTIER Version 4.1: A Computer
      Program for Stochastic Frontier Production and Cost Function
      Estimation, CEPA Working Paper 96/08,
      \url{http://www.uq.edu.au/economics/cepa/frontier.htm},
      University of New England.
}

\keyword{datasets}
