\name{telecom}
\docType{data}
\alias{telecom}

\title{Telecommunications Providers}

\description{
   The \code{telecom} data frame contains data
   on telecommunications providers in 21 countries in 1990.
}

\usage{data( telecom )}

\format{
   This data frame contains the following variables (columns):
   \describe{
      \item{country}{The name of the country.}
      \item{output}{Output (index).}
      \item{mainlines}{Mainlines (M km).}
      \item{employees}{Number of employees (10,000 persons).}
   }
}

\source{
Supplementary files for Coelli et al. (1998), p. 193.
}

\references{
Coelli, T. J., Rao, D. S. P., and Battese, G. E. (1998)
\emph{An Introduction to Efficiency and Productivity Analysis},
Springer, New York.
}

\keyword{datasets}
